Refactor existing PR
- remove userRole-Method
- set existing userRole-Pipe global and use it in the actionbar
- set no user-select on cards to prevent selecting parts of it
- change link to button to prevent navigating if it is not allowed
